Avoid resetting access policies when no access policy is defined
When deploying the building block with no access policy set,  all existing access policies are removed.
This leads to unwanted behaviour when deploying the building block as part of shared infrastructure,
especially when app-specific templates that create resources with managed identities are used.

Note that if an access policy is supplied to the building block, all other access policies not in the building block are still removed